Website Finished
I have done the css for every website, and changed every button to one of two formats, I've also changed the way the delete account function works as before there was no checker to see if the account had been logged into first before deleting so the user could just type the link with a certain account name and it would delete the account so I've added a checker. I've also gone through and added a checker for every function to check whether or not their account is logged into and whether or not the task id number is available as well and just redirecting them back to the to do list page which would further redirect them to the login page if not logged in.